Optimizing bootstrap download, when bootstrapping multiple nodes.

+++ dogecash_check_resync_all.sh
@@ -164,17 +164,21 @@ for FILE in $(ls ~/bin/${NAME}d_$ALIAS.sh | sort -V); do
 
   if [ -z "$NODEPID" ] && [ "$BOOTSTRAPURL" ]; then
     echo "Node $NODEALIAS is STOPPED"
+
     # Create Temp folder
     mkdir -p $CONF_DIR_TMP
     cd $CONF_DIR_TMP
 
-    echo "Downloading bootstrap"
-    if [[ $BOOTSTRAPURL == *.tar.gz ]]; then
+    if [ ! -f "bootstrap.tar.gz" ] && [[ $BOOTSTRAPURL == *.tar.gz ]]; then
+      echo "Downloading bootstrap"
       wget ${BOOTSTRAPURL} -O bootstrap.tar.gz
       WGET=$?
-    elif [[ $BOOTSTRAPURL == *.zip ]]; then
+    elif [ ! -f "bootstrap.zip" ] && [[ $BOOTSTRAPURL == *.zip ]]; then
+      echo "Downloading bootstrap"
       wget ${BOOTSTRAPURL} -O bootstrap.zip
       WGET=$?
+    else
+      echo "Bootstrap already exists, skipping download"
     fi
 
     if [ $WGET -eq 0 ]; then
@@ -220,3 +224,8 @@ for FILE in $(ls ~/bin/${NAME}d_$ALIAS.sh | sort -V); do
   fi
   done
 done
+
+# Remove temp folder
+if [ -d "$CONF_DIR_TMP" ]; then
+  rm -rfd $CONF_DIR_TMP
+fi
